requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
pycparser==3.0
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2025.3
psycopg2-binary==2.9.9
reportlab==4.1.0
firebase-admin==6.5.0
